sacheon/requirements.txt
azure-ai-textanalytics==5.3.0
azure-core==1.30.0
pandas==2.1.0
python-dotenv==1.0.0 